Reduce demo page layout jump
This reserves some vertical space for the NASA APOD image so the
page doesn't jump so much once the image loads. The height is just
a guess though, because the aspect ratio of the picture can change
by the day. I could probably do better auto-fitting with some fancy
css, but for now I suppose this is good enough.